pyproject.toml
[project]
name = "bkw-design-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.8.0",
    "anthropic>=0.20.0",
    "eventlet>=0.33.0",
    "fastmcp>=2.8.1",
    "flask>=3.0.0",
    "flask-mail>=0.9.1",
    "flask-socketio>=5.3.0",
    "googlemaps>=4.10.0",
    "groq>=0.4.0",
    "ipykernel>=7.0.1",
    "joblib>=1.3.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "langgraph>=1.0.0",
    "mcp>=1.0.0",
    "mcp-server<=0.1.4",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",
    "protobuf>=3.20.0",
    "python-dotenv>=1.0.0",
    "python-engineio>=4.7.0",
    "python-socketio>=5.9.0",
    "requests>=2.32.5",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0,<1.15.0",
    "setuptools>=80.9.0",
    "tavily-python>=0.7.12",
]
